+++ src/main/java/au/com/sealink/quicktravel/client/models/barcodes/core/TicketTemplate.java
@@ -44,6 +44,10 @@ public class TicketTemplate {
         return maximumUses;
     }
 
+    public void setMaximumUses(int maximumUses) {
+        this.maximumUses = maximumUses;
+    }
+
     public int getValidityDurationMinutes() {
         return validityDurationMinutes;
     }